Tblue-1.8 be-fix-for-icident-view

+++ src/main/java/ru/ssau/citizen/repository/PhotoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.ssau.citizen.entities.Photo;
 @Repository
-public interface PhotoRepository extends JpaRepository<Photo, Long> {
+public interface PhotoRepository extends JpaRepository<Photo, Integer> {
 }
